floats

+++ Part2A.C
@@ -227,15 +227,15 @@ public:
 		else{
 			TH1F *pbglPlot = new TH1F(name.c_str(),"",100,0,20);
 
-			double cerenkovEnergies[3]; // only need the [1] value 
+			float cerenkovEnergies[3]; // only need the [1] value 
 	 		orange->SetBranchAddress("TOWER_CALIB_C2.energy", &cerenkovEnergies);
-	 		double vetoEnergy[4]; //need all the values 
+	 		float vetoEnergy[4]; //need all the values 
 	 		orange->SetBranchAddress("TOWER_CALIB_TRIGGER_VETO.energy", &vetoEnergy);
-	 		double hodoVerticalEnergy[8];
+	 		float hodoVerticalEnergy[8];
 			orange->SetBranchAddress("TOWER_CALIB_HODO_VERTICAL.energy", &hodoVerticalEnergy);
-			double hodoHorizontalEnergy[8];
+			float hodoHorizontalEnergy[8];
 			orange->SetBranchAddress("TOWER_CALIB_HODO_HORIZONTAL.energy", &hodoHorizontalEnergy);
-			double pbglTemp;
+			float pbglTemp;
 			orange->SetBranchAddress("TOWER_CALIB_PbGL.energy", &pbglTemp);
 			for (int i = 0; i < SIZE; ++i)
 			{
@@ -252,14 +252,14 @@ public:
 		}
 	}
 	~OfficalBeamData(){}
-	inline bool passCuts(double cerenkov, double* veto, double* vhodo, double* hhodo){
+	inline bool passCuts(float cerenkov, float* veto, float* vhodo, float* hhodo){
 	//	return cerenkov>CERENKOVcut && noVeto(veto),passHodo(vhodo),passHodo(hhodo);
 		return true;
 	}
-	inline bool noVeto(double* veto){
+	inline bool noVeto(float* veto){
 		return veto[0]<VETOcut && veto[1]<VETOcut && veto[2]<VETOcut && veto[3]<VETOcut;
 	}
-	inline bool passHodo(double* hodo){
+	inline bool passHodo(float* hodo){
 		return hodo[0]>HODOcut && hodo[1]>HODOcut && hodo[2]>HODOcut && hodo[3]>HODOcut && hodo[4]>HODOcut && hodo[5]>HODOcut && hodo[6]>HODOcut && hodo[7]>HODOcut;
 	}
 	OfficalBeamData& operator=(OfficalBeamData other){
